Initial commit: first test not passing (TDD: 'red' phase)
I was having initial problems to get MonoDevelop/VS4Mac to show
up the NUnit tests in the unit-test panel (this must be a bug
that only affects .NETCore+NUnit libs), so what I went for in
the end is:

- FSharpExchangeDotNetStandard: .NET Standard lib with the core
functionality of the exchange.
- FSharpExchangeDotNetCoreApp: .NET Core App that can run the
code from the previous library (.NET Core <-> .NET Standard
interoperability).
- FSharpExchange.Tests: .NET 4.x library with NUnit 2.4.6 dep
that can unit test the FSharpExchangeDotNetStandard logic
(.NET Standard <-> .NET 4.x interoperability).

Not working:
- FSharpExhangeDotNetCore.Tests: we leave this skeleton here
to see if in the future MonoDevelop/VS4Mac can fix this bug.