FsharpExchange: store orders "in order", so the head is the tip (green)
To make this test pass, another less-efficient approach could
have done first, such as implementing a GetTip() method on
the OrderBookSide type whenever we want to see if an incoming
order matches or not. However, when I started to implement this
I found it really tricky given that in the case there's a match,
we want to remove the tip order (and replace it with another one
with less quantity, if it's a partial fill), and removing from
the list is again much less efficient, but most importantly,
tricky to implement given that there's no "remove" method in
F#'s List type[1].

However, by assuming a certain order at append-time, then we
can easily remove the tip when exploring the list in a functional
head::tail fashion very easily.

[1] https://stackoverflow.com/questions/48204224/why-fs-list-type-doesnt-have-a-remove-function