MakerOnly orders: refactor types to allow them, add basic test
The basic test added is just sending a buy MakerOnly order,
and a sell one, to an empty exchange, and asserting that
the exchange kept the order.

This doesn't cover yet the key difference between a normal
limit order and a MakerOnly limit order (this will be
covered in tests to be included in subsequent commits), so
the tests are still green.

This commit is mainly a refactoring of the Exchange API to
accept these kind of orders, but behaviour doesn't change.