Better naming and performance
I wanted to avoid a JSON serialization everytime an
orderBookSide fragment was created, and this achieves
it. To do it I had to create a new OrderBookSide type,
but as that name was already taken, I renamed the
previous type that had this name to 'OrderBookSideFragment'
because it can represent parts of an orderBookSide, not
just the whole of it.