Prevent order with same ID to be inserted
Even if we should assume that anything that gets sent to
the matching engine is in good faith (it should be another
service the one that should check for balances and whatnot),
it could happen that due to a bug, same order tries to get
inserted twice. The problem about this is that, as demonstrated
by the MakerOnly test I'm fixing in this commit, this could
have bad consequences: an infinite loop in the matching engine.

To prevent this from happening we now code a sanitycheck of
the order before it tries to get inserted: it checks if it
already exists. Maybe this check is too expensive for PROD
and should only be enabled in UAT/TEST environments... but
for now I leave it here just in case. FAIL FAST!